Conflicts (fixed):
	drivers/wireless/ieee802154/mrf24j40.c
	include/nuttx/fs/ioctl.h

+++ libc/math/lib_log10f.c
@@ -6,7 +6,7 @@
  *   Copyright (C) 2012 Gregory Nutt. All rights reserved.
  *   Ported by: Darcy Gong
  *
- * It derives from the Rhombs OS math library by Nick Johnson which has
+ * It derives from the Rhombus OS math library by Nick Johnson which has
  * a compatibile, MIT-style license:
  *
  * Copyright (C) 2009, 2010 Nick Johnson <nickbjohnson4224 at gmail.com>
@@ -37,5 +37,5 @@
 
 float log10f(float x)
 {
-  return (logf(x) / M_LN10);
+  return (logf(x) / (float)M_LN10);
 }
